Fix and split up tests
