+++ src/protocol/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Protocol tests.
+
+use super::*;
+
+mod validate_nonce;
+mod signaling_messages;
